script: Use `RequestGenerationId` for all stylesheet loads
Instead of just using the `RequestGenerationId` for the original load of
a `<link>` stylsheet, use it as well for all subsequent `@import`
stylsheets as well. When an element has a stylesheet with an `@import`
and then the element changes, those subsequent `@import`s should also be
ignored. That means that they should not not delay triggering of `load`
and `error` events and we should not waste time parsing them.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>